edit cleaning code and update data

+++ data_cleaning.py
@@ -18,7 +18,8 @@ ds['TeamExp'].fillna(int(mean_TeamExp), inplace = True)
 ds['ManagerExp'].fillna(int(mean_ManagerExp), inplace = True)
 
 #language field
-lang = ds['Language'].apply(lambda x: int(x.replace("b'","").replace("'","")))
+lang = ds['Language'].apply(lambda x: float(x.replace("b'","").replace("'","")))
 ds['Language'] = lang
 
 ds.to_csv('desharnais_clean.csv', index = False)
+
